feat: Add disord_bing example
