+++ .env.example
@@ -35,6 +35,7 @@ REDIS_PORT=
 #dbs
 PINECONE_API_KEY=""
 BING_COOKIE=""
+BING_AUTH=""
 
 # RevGpt Configuration
 ACCESS_TOKEN="your_access_token_here"
