+++ .gitignore
@@ -47,6 +47,8 @@ share/python-wheels/
 .installed.cfg
 *.egg
 MANIFEST
+output/*
+cookes.json
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
